feat: i like curry
